historical_data_pipeline/dashboard/requirements.txt
boto3
openpyxl
pandas
redshift_connector
python-dotenv
pylint
altair
pyarrow
fastparquet
streamlit